sendBack/Reject Advance Fix
