+++ src/app/fyle/view-team-advance/advance-actions/advance-actions.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { PopoverController } from '@ionic/angular';
 import { ApproveAdvanceComponent } from '../approve-advance/approve-advance.component';
-import { SendBackAdvanceComponent } from '../send-back-advance/send-back-advance.component';
-import { RejectAdvanceComponent } from '../reject-advance/reject-advance.component';
 
 @Component({
   selector: 'app-advance-actions',
