--- src/app/fyle/view-team-advance/reject-advance/reject-advance.component.html
@@ -1,63 +1,0 @@
-<ion-header>
-  <ion-toolbar class="reject--header">
-    <ion-title class="reject--header-text">Reject</ion-title>
-    <ion-buttons slot="end">
-      <ion-button (click)="cancel()" class="reject--header-cancel"> Cancel </ion-button>
-    </ion-buttons>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-  <div class="reject">
-    <div class="reject--body">
-      <div class="reject--card">
-        <div class="reject--value">
-          <span class="reject--card-currency">
-            {{ areq.areq_currency }}
-          </span>
-          <span class="reject--card-amount">
-            {{ areq.areq_amount }}
-          </span>
-        </div>
-        <div>
-          Submitted by <span>{{ areq.us_full_name }}</span>
-        </div>
-        <div>
-          <div>
-            Submitted on <span>{{ areq.areq_created_at | date: 'MMM dd, yyyy' }}</span>
-          </div>
-        </div>
-      </div>
-
-      <div class="reject--card">
-        <p>Please mention the reason for rejecting the advance request *</p>
-        <textarea
-          class="reject--textarea"
-          name="reason"
-          [(ngModel)]="rejectReason"
-          required
-          #textInput="ngModel"
-        ></textarea>
-        <div class="reject--validation" *ngIf="showReasonError">Please enter a reason.</div>
-      </div>
-    </div>
-  </div>
-</ion-content>
-
-<ion-footer>
-  <ion-toolbar>
-    <div class="reject--buttons">
-      <button
-        class="reject--button"
-        mat-flat-button
-        color="primary"
-        (click)="reject($event)"
-        appFormButtonValidation
-        [loading]="rejectLoading"
-        [loadingText]="'REJECTING'"
-      >
-        REJECT ADVANCE
-      </button>
-    </div>
-  </ion-toolbar>
-</ion-footer>
