--- src/app/fyle/view-team-advance/reject-advance/reject-advance.component.scss
@@ -1,130 +1,0 @@
-$details-color: #ababab;
-
-.reject {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: space-between;
-
-  &--header {
-    padding-top: 16px;
-    display: flex;
-    justify-content: space-between;
-  }
-
-  &--header-text {
-    font-size: 24px;
-    line-height: 1.3;
-    font-weight: 700;
-  }
-
-  &--header-cancel {
-    color: #0077ee;
-    font-weight: 400;
-    font-size: 16px;
-    text-transform: capitalize;
-  }
-
-  &--body {
-    height: 100%;
-    color: #ababab;
-    font-size: 16px;
-    font-weight: 500;
-    line-height: 1.5;
-    padding: 8px 24px;
-  }
-
-  &--buttons {
-    display: flex;
-    justify-content: center;
-  }
-
-  &--button {
-    width: 100%;
-    margin: 12px;
-    height: 48px;
-    font-weight: 600;
-  }
-
-  &--primary-cta {
-    padding: 16px;
-    .mat-button-base {
-      width: 100%;
-      font-weight: 700;
-      min-height: 47px;
-    }
-  }
-
-  &--redirect {
-    text-decoration: none;
-  }
-
-  &--card {
-    box-shadow: 0 0 8px rgba(0, 0, 0, 0.1);
-    border: 1px solid #e5e5e5;
-    padding: 12px;
-    border-radius: 4px;
-    &:not(:last-child) {
-      margin-bottom: 12px;
-    }
-  }
-
-  &--card-currency {
-    color: #ababab;
-    font-size: 16px;
-  }
-
-  &--card-amount {
-    color: #000;
-    font-size: 20px;
-  }
-
-  &--card-purpose {
-    line-height: 1.5;
-    color: #4a4a4a;
-    margin: 4px 0;
-    font-size: 16px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    border-bottom: 1px solid #efefef;
-    padding-bottom: 8px;
-  }
-
-  &--card-count {
-    font-size: 16px;
-    color: #ababab;
-    line-height: 1.5;
-  }
-  &--violations {
-    background-color: rgba(255, 153, 0, 0.25);
-    color: #ff9900;
-    font-size: 16px;
-    padding: 8px 16px;
-    margin-top: 16px;
-    border-radius: 4px;
-  }
-
-  &--textarea {
-    display: block;
-    width: 100%;
-    border: 1px solid #e0e0e0;
-    color: #000;
-    font-weight: normal;
-    margin-top: 8px;
-    height: 75px;
-    border-radius: 4px;
-    resize: none;
-  }
-
-  &--validation {
-    color: red;
-  }
-
-  &--reason {
-    border: 1px solid #efefef;
-    padding: 8px;
-    margin-top: 12px;
-    box-shadow: 0 0 8px rgba(0, 0, 0, 0.1);
-  }
-}
