--- src/app/fyle/view-team-advance/reject-advance/reject-advance.component.spec.ts
@@ -1,26 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { RejectAdvanceComponent } from './reject-advance.component';
-
-describe('RejectAdvanceComponent', () => {
-  let component: RejectAdvanceComponent;
-  let fixture: ComponentFixture<RejectAdvanceComponent>;
-
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        declarations: [RejectAdvanceComponent],
-        imports: [IonicModule.forRoot()],
-      }).compileComponents();
-
-      fixture = TestBed.createComponent(RejectAdvanceComponent);
-      component = fixture.componentInstance;
-      fixture.detectChanges();
-    })
-  );
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
