--- src/app/fyle/view-team-advance/send-back-advance/send-back-advance.component.scss
@@ -1,136 +1,0 @@
-$details-color: #ababab;
-
-.send-back {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: space-between;
-  &--header {
-    padding: 24px;
-    display: flex;
-    justify-content: space-between;
-  }
-
-  &--header-text {
-    font-size: 24px;
-    font-weight: 700;
-  }
-
-  &--header-cancel {
-    color: #0077ee;
-    font-weight: 400;
-    font-size: 16px;
-  }
-
-  &--body {
-    height: 100%;
-    color: #ababab;
-    font-size: 16px;
-    line-height: 1.5;
-    padding: 8px 24px;
-  }
-
-  &--buttons {
-    display: flex;
-    justify-content: center;
-  }
-
-  &--button {
-    height: 48px;
-    width: 100%;
-    margin: 4px 24px 12px;
-  }
-
-  &--primary-cta {
-    padding: 16px;
-    .mat-button-base {
-      width: 100%;
-      font-weight: 700;
-      min-height: 47px;
-    }
-  }
-
-  &--redirect {
-    text-decoration: none;
-  }
-
-  &--card {
-    box-shadow: 0 0 8px rgba(0, 0, 0, 0.1);
-    border: 1px solid #e5e5e5;
-    padding: 12px;
-    border-radius: 4px;
-    &:not(:last-child) {
-      margin-bottom: 12px;
-    }
-  }
-
-  &--card-currency {
-    color: #ababab;
-    font-size: 16px;
-  }
-
-  &--card-amount {
-    color: #000;
-    font-size: 20px;
-  }
-
-  &--card-purpose {
-    line-height: 1.5;
-    color: #4a4a4a;
-    margin: 4px 0;
-    font-size: 16px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    border-bottom: 1px solid #efefef;
-    padding-bottom: 8px;
-  }
-
-  &--card-count {
-    font-size: 16px;
-    color: #ababab;
-    line-height: 1.5;
-  }
-  &--violations {
-    background-color: rgba(255, 153, 0, 0.25);
-    color: #ff9900;
-    font-size: 16px;
-    padding: 8px 16px;
-    margin-top: 16px;
-    border-radius: 4px;
-  }
-
-  &--textarea {
-    display: block;
-    width: 100%;
-    border: 1px solid #e0e0e0;
-    margin-top: 8px;
-    height: 75px;
-    border-radius: 4px;
-    resize: none;
-  }
-
-  &--validation {
-    color: red;
-  }
-
-  &--reason {
-    border: 1px solid #efefef;
-    padding: 16px;
-    margin-top: 12px;
-    box-shadow: 0 0 8px rgba(0, 0, 0, 0.1);
-  }
-
-  &--date-container {
-    border-top: 1px solid #e0e0e0;
-    margin-top: 8px;
-    padding-top: 8px;
-  }
-
-  &--name-container {
-    line-height: 1.5;
-    color: #4a4a4a;
-    margin: 4px 0;
-    font-size: 16px;
-  }
-}
