--- src/app/fyle/view-team-advance/send-back-advance/send-back-advance.component.spec.ts
@@ -1,26 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { SendBackAdvanceComponent } from './send-back-advance.component';
-
-describe('SendBackAdvanceComponent', () => {
-  let component: SendBackAdvanceComponent;
-  let fixture: ComponentFixture<SendBackAdvanceComponent>;
-
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        declarations: [SendBackAdvanceComponent],
-        imports: [IonicModule.forRoot()],
-      }).compileComponents();
-
-      fixture = TestBed.createComponent(SendBackAdvanceComponent);
-      component = fixture.componentInstance;
-      fixture.detectChanges();
-    })
-  );
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
