+++ src/app/fyle/view-team-advance/view-team-advance.module.ts
@@ -9,9 +9,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatRippleModule } from '@angular/material/core';
 import { SharedModule } from '../../shared/shared.module';
 import { AdvanceActionsComponent } from './advance-actions/advance-actions.component';
-import { RejectAdvanceComponent } from './reject-advance/reject-advance.component';
 import { ApproveAdvanceComponent } from './approve-advance/approve-advance.component';
-import { SendBackAdvanceComponent } from './send-back-advance/send-back-advance.component';
 
 @NgModule({
   imports: [
@@ -24,12 +22,6 @@ import { SendBackAdvanceComponent } from './send-back-advance/send-back-advance.
     MatRippleModule,
     SharedModule,
   ],
-  declarations: [
-    ViewTeamAdvancePage,
-    AdvanceActionsComponent,
-    ApproveAdvanceComponent,
-    RejectAdvanceComponent,
-    SendBackAdvanceComponent,
-  ],
+  declarations: [ViewTeamAdvancePage, AdvanceActionsComponent, ApproveAdvanceComponent],
 })
 export class ViewTeamAdvancePageModule {}
